requirements.txt
# Django
Django==3.2.16

# Rest Framewrok
djangorestframework==3.14.0

# Swagger dependency
drf-yasg==1.21.4

# Pytest 
pytest-django==4.5.2
pytest==7.2.0
factory-boy==3.2.1

# deploy 
gunicorn==20.1.0

# Environment
python-dotenv==0.21.0

